Rename some headers.
