+++ example/bencher.cpp
@@ -1,6 +1,6 @@
-#include "arviss/common/types.h"         // Address
-#include "arviss/platforms/basic/cpus.h" // BasicRv32imfCpu
-#include "arviss/rv32/concepts.h"        // IsRv32imfCpu
+#include "arviss/common/types.h"          // Address
+#include "arviss/platforms/basic/basic.h" // BasicRv32imfCpu
+#include "arviss/rv32/concepts.h"         // IsRv32imfCpu
 
 #include <format>
 #include <fstream>
